pyproject.toml
[tool.poetry]
name = "rdo-helper"
version = "0.3.0"
description = "A Discord bot for Red Dead Online that provides daily challenge guides and troubleshooting help."
authors = ["Iris-Raine <iris.raine.darkest.star@gmail.com>"]
readme = "README.md"
packages = [
  {include = "engine"}
]

[tool.poetry.dependencies]
python = "^3.10"
nextcord = "^2.6.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[tool.poetry.scripts]
rdo-helper-bot = "engine.run:run_discord_bot"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
